feat: refactor commands/verify.py into verify/ package
Split 829-line verify.py into 6 focused modules following brainstorm/digest pattern.

**Structure:**
- __init__.py (66 lines) - CLI entry point
- citation_verifier.py (112 lines) - Citation verification
- reasoning_handler.py (251 lines) - Reasoning trace operations
- soundness_checker.py (178 lines) - Legal soundness validation
- core.py (333 lines) - Main orchestration & CoVe
- formatters.py (93 lines) - Report formatting helpers

**Changes:**
- Updated 13 test patches in test_verify_command.py
- All 387 tests passing, ruff clean
- Zero breaking changes to CLI or public API
