+++ tests/unit/test_verify_command.py
@@ -73,10 +73,13 @@ class TestVerifyCommand:
             f.write(sample_legal_text)
 
         with (
-            patch("litassist.commands.verify.verify_all_citations") as mock_citations,
-            patch("litassist.commands.verify.fetch_citation_context") as mock_fetch,
-            patch("litassist.commands.verify.LLMClientFactory") as mock_llm_factory,
-            patch("litassist.commands.verify.save_log") as _mock_save_log,
+            patch("litassist.commands.verify.citation_verifier.verify_all_citations") as mock_citations,
+            patch("litassist.commands.verify.citation_verifier.fetch_citation_context") as mock_fetch,
+            patch("litassist.commands.verify.reasoning_handler.fetch_citation_context") as mock_reasoning_fetch,
+            patch("litassist.commands.verify.soundness_checker.fetch_citation_context") as mock_soundness_fetch,
+            patch("litassist.commands.verify.reasoning_handler.LLMClientFactory") as mock_reasoning_factory,
+            patch("litassist.commands.verify.soundness_checker.LLMClientFactory") as mock_soundness_factory,
+            patch("litassist.commands.verify.core.save_log") as _mock_save_log,
         ):
             mock_citations.return_value = (
                 [
@@ -86,13 +89,16 @@ class TestVerifyCommand:
                 [("Smith v Jones [2025] NSWSC 999", "Future citation")],
             )
             mock_fetch.return_value = {}  # Mock empty case content
+            mock_reasoning_fetch.return_value = {}  # Mock empty case content
+            mock_soundness_fetch.return_value = {}  # Mock empty case content
             mock_client = Mock()
             mock_client.verify.return_value = (
                 "No legal issues found.",
                 "anthropic/claude-opus-4.1",
             )
             mock_client.complete.return_value = ("Analysis with reasoning trace", {})
-            mock_llm_factory.for_command.return_value = mock_client
+            mock_reasoning_factory.for_command.return_value = mock_client
+            mock_soundness_factory.for_command.return_value = mock_client
 
             result = runner.invoke(verify, [temp_file])
             assert result.exit_code == 0
@@ -112,10 +118,10 @@ class TestVerifyCommand:
             f.write(sample_legal_text)
 
         with (
-            patch("litassist.commands.verify.verify_all_citations") as mock_citations,
-            patch("litassist.commands.verify.fetch_citation_context") as mock_fetch,
-            patch("litassist.commands.verify.extract_citations") as mock_extract,
-            patch("litassist.commands.verify.save_log") as _mock_save_log,
+            patch("litassist.commands.verify.citation_verifier.verify_all_citations") as mock_citations,
+            patch("litassist.commands.verify.citation_verifier.fetch_citation_context") as mock_fetch,
+            patch("litassist.commands.verify.citation_verifier.extract_citations") as mock_extract,
+            patch("litassist.commands.verify.core.save_log") as _mock_save_log,
         ):
             mock_extract.return_value = ["Mabo v Queensland (No 2) [1992] HCA 23"]
             mock_citations.return_value = (
@@ -137,9 +143,9 @@ class TestVerifyCommand:
             f.write(sample_legal_text)
 
         with (
-            patch("litassist.commands.verify.fetch_citation_context") as mock_fetch,
-            patch("litassist.commands.verify.LLMClientFactory") as mock_llm_factory,
-            patch("litassist.commands.verify.save_log") as _mock_save_log,
+            patch("litassist.commands.verify.soundness_checker.fetch_citation_context") as mock_fetch,
+            patch("litassist.commands.verify.soundness_checker.LLMClientFactory") as mock_llm_factory,
+            patch("litassist.commands.verify.core.save_log") as _mock_save_log,
         ):
             mock_fetch.return_value = {}  # Mock empty case content
             mock_client = Mock()
@@ -167,10 +173,12 @@ class TestVerifyCommand:
             f.write(sample_text_with_reasoning)
 
         with (
-            patch("litassist.commands.verify.fetch_citation_context") as mock_fetch,
-            patch("litassist.commands.verify.save_log") as _mock_save_log,
+            patch("litassist.commands.verify.soundness_checker.fetch_citation_context") as mock_fetch_soundness,
+            patch("litassist.commands.verify.reasoning_handler.fetch_citation_context") as mock_fetch_reasoning,
+            patch("litassist.commands.verify.core.save_log") as _mock_save_log,
         ):
-            mock_fetch.return_value = {}  # Mock empty case content
+            mock_fetch_soundness.return_value = {}  # Mock empty case content
+            mock_fetch_reasoning.return_value = {}  # Mock empty case content
             result = runner.invoke(verify, [temp_file, "--reasoning"])
             assert result.exit_code == 0
             assert "Reasoning trace verified" in result.output
@@ -185,9 +193,9 @@ class TestVerifyCommand:
             f.write(sample_legal_text)
 
         with (
-            patch("litassist.commands.verify.fetch_citation_context") as mock_fetch,
-            patch("litassist.commands.verify.LLMClientFactory") as mock_llm_factory,
-            patch("litassist.commands.verify.save_log") as _mock_save_log,
+            patch("litassist.commands.verify.reasoning_handler.fetch_citation_context") as mock_fetch,
+            patch("litassist.commands.verify.reasoning_handler.LLMClientFactory") as mock_llm_factory,
+            patch("litassist.commands.verify.core.save_log") as _mock_save_log,
         ):
             mock_fetch.return_value = {}  # Mock empty case content
             mock_client = Mock()
@@ -297,7 +305,7 @@ class TestVerifyCommand:
         """Test handling of API failures."""
         with open(temp_file, "w") as f:
             f.write(sample_legal_text)
-        with patch("litassist.commands.verify.verify_all_citations") as mock_citations:
+        with patch("litassist.commands.verify.citation_verifier.verify_all_citations") as mock_citations:
             mock_citations.side_effect = Exception("API unavailable")
             result = runner.invoke(verify, [temp_file, "--citations"])
             assert "Citation verification failed" in result.output
@@ -308,17 +316,23 @@ class TestVerifyCommand:
         with open(temp_file, "w") as f:
             f.write(sample_legal_text)
         with (
-            patch("litassist.commands.verify.verify_all_citations") as mock_citations,
-            patch("litassist.commands.verify.fetch_citation_context") as mock_fetch,
-            patch("litassist.commands.verify.LLMClientFactory") as mock_llm_factory,
-            patch("litassist.commands.verify.save_log") as _mock_save_log,
+            patch("litassist.commands.verify.citation_verifier.verify_all_citations") as mock_citations,
+            patch("litassist.commands.verify.citation_verifier.fetch_citation_context") as mock_fetch,
+            patch("litassist.commands.verify.reasoning_handler.fetch_citation_context") as mock_reasoning_fetch,
+            patch("litassist.commands.verify.soundness_checker.fetch_citation_context") as mock_soundness_fetch,
+            patch("litassist.commands.verify.reasoning_handler.LLMClientFactory") as mock_reasoning_factory,
+            patch("litassist.commands.verify.soundness_checker.LLMClientFactory") as mock_soundness_factory,
+            patch("litassist.commands.verify.core.save_log") as _mock_save_log,
         ):
             mock_citations.return_value = (["Case1"], [])
             mock_fetch.return_value = {}  # Mock empty case content
+            mock_reasoning_fetch.return_value = {}  # Mock empty case content
+            mock_soundness_fetch.return_value = {}  # Mock empty case content
             mock_client = Mock()
             mock_client.verify.return_value = ("No issues", "anthropic/claude-opus-4.1")
             mock_client.complete.return_value = ("Analysis", {})
-            mock_llm_factory.for_command.return_value = mock_client
+            mock_reasoning_factory.for_command.return_value = mock_client
+            mock_soundness_factory.for_command.return_value = mock_client
             result = runner.invoke(verify, [temp_file])
             assert result.exit_code == 0
             assert "_citations_" in result.output
